Deal now has performPlay
Massive changes to Play -- perform obviously moved out, some functions moved out
of .hpp, Trick now actually takes ownership of the cards it gets (they were actually leaking
earlier!) and therefore is not copyable anymore. Play now keeps a history in the
form of a list of Tricks. The equality operator for Suit and Denomination seems
to be dangerous, but I only moved it, so no changes there.
One more change to Deal -- bidding and play are now unique pointers and are
initiated when the proper function is called. This is not only neccessary for
Play, but will also allow for calling update() when implementing Observable properly.