Handling dummy role
Roles are assigned to arbiters in Deal::performPlay().
Apart from the reference to the actual player, every arbiter holds now a reference to the player's partner to be able to handle the dummy role and forward the request to the partner. Since the partner's chooseCardFromDummy method takes the partner's hand as well, it should be set by arbiter's setPartnerHand right aftersetting arbiter's role to DUMMY.